backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
flask-swagger-ui